Update populators and prepare cherrytwist-populator for publish
