+++ .npmignore
@@ -1,0 +1,4 @@
+.github
+src
+tsconfig.json
+.vscode
