+++ src/populator.ts
@@ -1,6 +1,7 @@
 import { CherrytwistClient } from 'cherrytwist-lib';
 import { Logger } from 'winston';
-import { DataAdapter } from './adapters/adapter';
+import { AbstractDataAdapter } from './adapters/data-adapter';
+import { AbstractPopulator } from './populators/abstract-populator';
 import { ChallengePopulator } from './populators/challenge-populator';
 import { EcoversePopulator } from './populators/ecoverse-populator';
 import { GroupPopulator } from './populators/group-populator';
@@ -9,30 +10,15 @@ import { OpportunityPopulator } from './populators/opportunity-populator';
 import { OrganisationPopulator } from './populators/organisation-populator';
 import { UserPopulator } from './populators/user-populator';
 
-export class GSheetParams {
-  gsheetID = '';
-  google_credentials_file = '';
-  google_token_file = '';
-}
-
-export class Populator {
-  // The ctClient to use to interact with the server
-  client: CherrytwistClient;
-  data: DataAdapter;
-  logger: Logger;
-  profiler: Logger;
-
+export class Populator extends AbstractPopulator {
   // Create the ecoverse with enough defaults set/ members populated
   constructor(
     client: CherrytwistClient,
-    data: DataAdapter,
+    data: AbstractDataAdapter,
     logger: Logger,
     profiler: Logger
   ) {
-    this.client = client;
-    this.data = data;
-    this.logger = logger;
-    this.profiler = profiler;
+    super(client, data, logger, profiler);
   }
 
   async populate() {
