+++ src/populators/abstract-populator.ts
@@ -1,17 +1,18 @@
 import { CherrytwistClient } from 'cherrytwist-lib';
 import { Logger } from 'winston';
-import { DataAdapter } from '../adapters/adapter';
+import { AbstractDataAdapter } from '../adapters/data-adapter';
 
 export abstract class AbstractPopulator {
-  client: CherrytwistClient;
-  data: DataAdapter;
-  logger: Logger;
-  profiler: Logger;
+  protected client: CherrytwistClient;
+  protected data: AbstractDataAdapter;
+  protected logger: Logger;
+  protected profiler: Logger;
+  protected name: string;
 
   // Create the ecoverse with enough defaults set/ members populated
   constructor(
     client: CherrytwistClient,
-    data: DataAdapter,
+    data: AbstractDataAdapter,
     logger: Logger,
     profiler: Logger
   ) {
@@ -19,6 +20,7 @@ export abstract class AbstractPopulator {
     this.data = data;
     this.logger = logger;
     this.profiler = profiler;
+    this.name = 'abstract-populator';
   }
 
   abstract populate(): void;
