+++ src/populators/challenge-populator.ts
@@ -1,6 +1,6 @@
 import { CherrytwistClient, Organisation } from 'cherrytwist-lib';
 import { Logger } from 'winston';
-import { DataAdapter } from '../adapters/adapter';
+import { AbstractDataAdapter } from '../adapters/data-adapter';
 import { Challenge } from '../models';
 import { AbstractPopulator } from './abstract-populator';
 
@@ -10,7 +10,7 @@ export class ChallengePopulator extends AbstractPopulator {
 
   constructor(
     client: CherrytwistClient,
-    data: DataAdapter,
+    data: AbstractDataAdapter,
     logger: Logger,
     profiler: Logger
   ) {
