+++ src/populators/index.ts
@@ -1,0 +1,8 @@
+export * from './abstract-populator';
+export * from './challenge-populator';
+export * from './ecoverse-populator';
+export * from './group-populator';
+export * from './host-populator';
+export * from './opportunity-populator';
+export * from './organisation-populator';
+export * from './user-populator';
